added page caching to speed up performance on subsequent requests
not sure if i want to keep this in here or not, but downloading all the images is really slow right now, and if you're fiddling with the different pdf options it's a pain to wait for it to run every time. maybe a pool of threads to prefetch all the images would work, but might not be any faster with ruby's threading system. it's intriguing enough that i'll probably try it though.